LayoutTests/platform/wpe/fast/scrolling/sync-scroll-overscroll-behavior-unscrollable-iframe-expected.txt



FAIL Unscrollable iframe horizontal scroll test with overscroll-behavior: none. assert_equals: Propagate horizontal scroll expected true but got false
PASS Unscrollable iframe vertical scroll test with overscroll-behavior: none.
